third_party/WebKit/LayoutTests/editing/inserting/insert-on-unfocused-element.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="outer" tabindex="-1">
  <div id="input" contenteditable>foo</div>
</div>
<script>
test(() => {
  assert_own_property(window, 'eventSender', 'this test requires window.eventSender');
  input.focus();
  assert_equals(document.activeElement, input);
  outer.focus();
  assert_equals(document.activeElement, outer, 'outer is focued.');
  eventSender.keyDown('a');
  assert_equals(input.textContent, 'foo', 'Input value should not change.');
});
</script>
